+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next:
 
 - Documentation for Jenkins environment variables
+- Autocompletion for Jenkins environment variables
 
 ## [1.1.0](https://github.com/Maarti/JenkinsDocExtension/releases/tag/1.1.0) - 2021-03-10